Move object_cache to api_handle module
