+++ libvcx/src/api_lib/api_handle/credential_def.rs
@@ -6,7 +6,7 @@ use crate::libindy::utils::{anoncreds, ledger};
 use crate::libindy::utils::cache::update_rev_reg_ids_cache;
 use crate::libindy::utils::payments::PaymentTxn;
 use crate::utils::constants::DEFAULT_SERIALIZE_VERSION;
-use crate::utils::object_cache::ObjectCache;
+use crate::api_lib::api_handle::object_cache::ObjectCache;
 use crate::utils::serialization::ObjectWithVersion;
 
 lazy_static! {
